pyproject.toml
[project]
name = "rag-agentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.7",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.29",
    "langchain-openai>=1.1.13",
    "python-dotenv>=1.2.2",
]
